requirements.txt
google-genai
pinecone
python-dotenv
PyPDF2
opencv-python
pillow
unstructured
